fix enum circular import
