+++ insight-be/src/modules/timbuktu/administrative/style/page-element-type.ts
@@ -1,0 +1,11 @@
+import { registerEnumType } from '@nestjs/graphql';
+
+export enum PageElementType {
+  Text = 'text',
+  Table = 'table',
+  Image = 'image',
+  Video = 'video',
+  Quiz = 'quiz',
+}
+
+registerEnumType(PageElementType, { name: 'PageElementType' });
